pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thunder-forge"
version = "0.1.0"
description = "Minimal controller for a small self-hosted Ollama node fleet"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Shared Goals" }]
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "jinja2>=3.1",
  "pyyaml>=6.0",
  "python-multipart>=0.0.9",
  "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "pytest-cov>=5.0",
]

[project.scripts]
thunder-forge = "thunder_forge.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"thunder_forge" = ["templates/*.html"]
